Complete rewrite of how steppers are handled.
Stiff and nonstiff steppers (i.e., internal state) are both handled
in same object type.  Collection is done with boost::any rather than
boost::variant, though we have to keep a copy of that header in the
package now.

Runtime dispatch is done with a big ugly series of switch statements,
but they can actually be changed to a different approach relatively
easily.

Code generation with cog to remove some of the error-prone duplication.